# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: HydroNexus Water Management API
  version: 1.0.0
  description: >
    Production-ready REST API specification for the HydroNexus Municipal Water Management System.
    Serves both the Officer Operations Portal (port 5173) and Citizen Portal (port 5174).
servers:
  - url: http://localhost:8000/api/v1
    description: Local development backend server
  - url: https://api.water.municipality.gov.in/api/v1
    description: Production API gateway
paths:
...
  /complaints:
    get:
      summary: List complaints
      description: Returns complaints filtered by type, ward, or status.
      tags: [Complaints]
      security:
        - BearerAuth: []
      parameters:
        - name: filter
          in: query
          schema:
            type: string
          description: Optional filter query (e.g. status or category)
      responses:
        '200':
          description: List of complaints
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Complaint'
    post:
      summary: Submit complaint
      description: Files a new citizen water grievance.
      tags: [Complaints]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewComplaintRequest'
      responses:
        '201':
          description: Complaint created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/Complaint'
  /complaints/{id}:
    get:
      summary: Get complaint detail (officer)
      description: >
        Officer-only complaint detail view. Roles: any officer role.
        Query shape and status vocabulary match packages/types Complaint.
      tags: [Complaints]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Complaint detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/Complaint'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Citizen tokens and other non-officer roles are rejected
        '404':
          description: Unknown complaint id
    patch:
      summary: Update complaint
      description: >
        Updates status, priority, or field assignment of a complaint.
        Roles: Admin, Supervisor, Engineer. KNOWN DRIFT: besides the
        packages/types statuses (Open, In Progress, Resolved) the endpoint
        also accepts Pending (stored as Open) and Escalated, matching the
        citizen-side vocabulary in openapi.yaml. See "Known drift" note.
      tags: [Complaints]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [Open, Pending, In Progress, Resolved, Escalated]
                priority:
                  type: string
                  enum: [Low, Medium, High, Critical]
                assigned:
                  type: string
      responses:
        '200':
          description: Complaint updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/Complaint'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Role lacks complaint-write permission
        '404':
          description: Unknown complaint id
        '422':
          description: Validation failed (unknown enum, empty patch)
  /complaints/{id}/status:
    patch:
      summary: Update complaint status (api-client alias)
      description: >
        Alias used by @water/api-client updateStatus. Identical service and
        audit trail as PATCH /complaints/{id}. Roles: Admin, Supervisor,
        Engineer.
      tags: [Complaints]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [Open, Pending, In Progress, Resolved, Escalated]
      responses:
        '200':
          description: Complaint status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/Complaint'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Role lacks complaint-write permission
        '404':
          description: Unknown complaint id
        '422':
          description: Unknown status value
  /complaints/{id}/assign:
    post:
      summary: Assign complaint to ward engineer (officer)
      description: >
        Assigns a complaint to an engineer and creates (or links) a work
        order. Roles: Admin, Supervisor.
      tags: [Complaints]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [engineer]
              properties:
                engineer: { type: string, example: 'Anil Jadhav' }
                workOrderId: { type: string }
                notes: { type: string }
      responses:
        '200':
          description: Complaint assigned
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      complaint:
                        $ref: '#/components/schemas/Complaint'
                      workOrder:
                        $ref: '#/components/schemas/WorkOrder'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Role lacks assign permission
        '404':
          description: Unknown complaint id
        '422':
          description: Missing engineer
...
  /admin/officers:
    get:
      summary: List municipal staff and field operator directory
      tags: [Admin]
      security:
        - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OfficerUser'
    post:
      summary: Create officer (officer)
      description: 'Roles: Admin. Seed password Password123!.'
      tags: [Admin]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, phone, role, department, zone]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                phone: { type: string }
                role: { type: string, enum: [Admin, Engineer, Supervisor, Operator] }
                department: { type: string }
                zone: { type: string }
      responses:
        '201':
          description: Officer created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/OfficerUser'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Non-Admin role rejected
        '409':
          description: Email already exists
        '422':
          description: Validation failed
...
  /work-orders:
    get:
      summary: List work orders (officer)
      description: 'Roles: any officer role. Filters ?status= ?ward= ?assignee=.'
      tags: [Complaints]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Work order list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkOrder'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Non-officer role rejected
    post:
      summary: Create work order (officer)
      description: 'Roles: Admin, Supervisor, Engineer.'
      tags: [Complaints]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, ward, assignedTo]
              properties:
                complaintId: { type: string }
                title: { type: string }
                ward: { type: string }
                assignedTo: { type: string }
                priority: { type: string, enum: [Low, Medium, High, Critical] }
                notes: { type: string }
      responses:
        '201':
          description: Work order created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/WorkOrder'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Role lacks work-order write permission
        '422':
          description: Validation failed
  /work-orders/{id}:
    get:
      summary: Work order detail with linked complaint (officer)
      description: 'Roles: any officer role.'
      tags: [Complaints]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Work order detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/WorkOrder'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Non-officer role rejected
        '404':
          description: Unknown work order id
    patch:
      summary: Update work order (officer)
      description: 'Roles: Admin, Supervisor, Engineer.'
      tags: [Complaints]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [Open, In Progress, On Hold, Completed, Cancelled] }
                assignedTo: { type: string }
                notes: { type: string }
      responses:
        '200':
          description: Work order updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/WorkOrder'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Role lacks work-order write permission
        '404':
          description: Unknown work order id
        '422':
          description: Validation failed
...
  /analytics/complaints:
    get:
      summary: Complaint analytics (officer)
      description: 'Roles: any officer role. Optional ?range= (default 6m).'
      tags: [Analytics]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Complaints by category
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Non-officer role rejected
...
  /admin/officers/{id}:
    patch:
      summary: Update officer (officer)
      description: 'Roles: Admin. Self-deactivation is rejected with 403.'
      tags: [Admin]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { type: string, enum: [Admin, Engineer, Supervisor, Operator] }
                department: { type: string }
                zone: { type: string }
                status: { type: string, enum: [Active, Inactive] }
      responses:
        '200':
          description: Officer updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/OfficerUser'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Non-Admin role, or self-deactivation attempt
        '404':
          description: Unknown officer id
        '422':
          description: Validation failed
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Authentication token missing, invalid, or expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiError:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          required: [message]
          properties:
            message: { type: string, example: 'Unauthorized access' }
            code: { type: string, example: 'AUTH_UNAUTHORIZED' }
            status: { type: integer, example: 401 }
            details: { type: object }
...
    Complaint:
      type: object
      required: [id, type, ward]
      properties:
        id: { type: string, example: 'CMP-2024-0891' }
        type: { type: string, example: 'Leakage' }
        citizen: { type: string, example: 'Ramesh Shinde' }
        ward: { type: string, example: 'Kasba Bawada' }
        address: { type: string, example: 'Plot 14, Sector 4, Kasba Bawada' }
        location: { type: string, example: 'Kasba Bawada Chowk' }
        phone: { type: string, example: '9876543210' }
        reported: { type: string, example: '11 Sep, 07:40' }
        date: { type: string, example: '11 Sep 2026' }
        assigned: { type: string, example: 'Anil Jadhav' }
        priority:
          type: string
          enum: [Low, Medium, High, Critical]
          example: Critical
        status:
          type: string
          enum: [Pending, In Progress, Resolved, Escalated]
          example: In Progress
        description: { type: string, example: 'Major pipeline burst near chowk' }
        updated: { type: string, example: '11 Sep, 09:00' }
        icon: { type: string, example: 'water_damage' }
    NewComplaintRequest:
      type: object
      required: [type, ward, description]
      properties:
        type: { type: string, example: 'Leakage' }
        ward: { type: string, example: 'Ward 12 - Rankala' }
        address: { type: string, example: 'Rankala Lake Road, Gate 3' }
        description: { type: string, example: 'Water overflowing from underground pipe' }
        priority: { type: string, enum: [Low, Medium, High, Critical], default: Medium }
...
    WorkOrder:
      type: object
      properties:
        id: { type: string, example: 'WO-2026-0101' }
        complaintId: { type: string, example: 'CMP-2024-0891' }
        title: { type: string, example: 'Isolate feeder valve 4B and repair burst main' }
        type: { type: string, enum: [Preventive, Corrective, Inspection], example: Corrective }
        ward: { type: string, example: 'Ward 07 - Kasba Bawada' }
        scheduledDate: { type: string, example: '12 Sep 2026' }
        status:
          type: string
          enum: [Pending, Assigned, In Progress, Completed, Cancelled]
          example: Assigned
        priority:
          type: string
          enum: [Low, Medium, High, Critical]
          example: Critical
        assignedTeam: { type: string, example: 'Unit A - Mechanical' }
        assignedTo: { type: string, example: 'Anil Jadhav' }
        notes: { type: string }
...
    OfficerUser:
      type: object
      properties:
        id: { type: string, example: 'USR-OFFICER-001' }
        name: { type: string, example: 'Suresh Patil' }
        email: { type: string, example: 'suresh.patil@kmcwater.gov.in' }
        phone: { type: string, example: '+91 98220 12345' }
        role: { type: string, enum: [Admin, Engineer, Supervisor, Operator], example: Admin }
        department: { type: string, example: 'Water Supply Operations' }
        zone: { type: string, example: 'Central Zone' }
        status: { type: string, enum: [Active, Inactive], example: Active }
        lastActive: { type: string, example: 'Just now' }
...